Upgrading to Greynoise v2 (breaking changes) #86
This work upgrade to the newer version of Greynoise API. For this to work, you will need to set up an API key.

The reason to change this was that the v1 of Greynoise API was getting old and the queries were sent through HTTP.

This PR renames the greyhouse.py file to gn.py because python was complaining about the imports.